+++ src/FlaUILibrary/keywords/__init__.py
@@ -8,7 +8,7 @@ from .keyboard import KeyboardKeywords
 from .screenshot import ScreenshotKeywords
 from .textbox import TextBoxKeywords
 from .window import WindowKeywords
-from .listview import ListViewKeywords
+from .grid import GridKeywords
 from .radiobutton import RadioButtonKeywords
 from .listbox import ListBoxKeywords
 from .tab import TabKeywords